refactor: split Win32 enum types
